feat(fe2): removed old design versions
