--- packages/frontend-2/components/viewer/resources/VersionCardDesignOption2.vue
@@ -1,95 +1,0 @@
-<template>
-  <button
-    :class="`group block w-full transition text-left rounded-md ${
-      clickable ? 'hover:bg-primary-muted' : 'cursor-default'
-    }`"
-    @click="handleClick"
-  >
-    <div class="flex items-center space-x-2">
-      <div class="bg-foundation w-24 h-24 shadow rounded-md">
-        <PreviewImage :preview-url="previewUrl" />
-      </div>
-      <div class="flex flex-col space-y-1">
-        <div class="space-x-1">
-          <div
-            v-if="showMetadata"
-            class="inline-block rounded-full px-2 text-xs bg-foundation-focus xxxtext-foreground-on-primary font-bold"
-          >
-            <!-- {{ isLatest ? 'Latest' : timeAgoCreatedAt }} -->
-          </div>
-          <div
-            v-if="isLatestVersion && showMetadata"
-            class="inline-block rounded-full px-2 text-xs bg-foundation-focus xxxtext-foreground-on-primary font-bold"
-          >
-            Loaded
-          </div>
-          <div class="inline-block rounded-full px-1 text-xs text-primary font-bold">
-            {{ version.sourceApplication }}
-          </div>
-        </div>
-        <div v-if="author" class="flex items-center space-x-2 overflow-hidden">
-          <UserAvatar :user="author" size="sm" />
-          <span class="text-sm font-bold tracking-tight text-foreground">
-            {{ author.name }}
-          </span>
-        </div>
-        <div class="text-xs text-foreground-2">{{ createdAt }}</div>
-        <div class="text-xs text-foreground-2 hidden">
-          {{ timeAgoCreatedAt }}
-        </div>
-        <div class="text-xs text-foreground-2 truncate break-all">
-          {{ version.message || 'no description' }}
-        </div>
-      </div>
-    </div>
-  </button>
-</template>
-<script setup lang="ts">
-import dayjs from 'dayjs'
-import { ViewerModelVersionCardItemFragment } from '~~/lib/common/generated/gql/graphql'
-import {
-  useInjectedViewer,
-  useResolvedViewerResources
-} from '~~/lib/viewer/composables/viewer'
-import { useGetPreviewUrl } from '~~/lib/viewer/helpers'
-
-const props = withDefaults(
-  defineProps<{
-    version: ViewerModelVersionCardItemFragment
-    showMetadata?: boolean
-    clickable?: boolean
-    modelId: string
-    isLatestVersion?: boolean
-  }>(),
-  {
-    showMetadata: true,
-    clickable: true
-  }
-)
-
-const emit = defineEmits<{
-  (e: 'change-version', version: string): void
-}>()
-
-const { projectId } = useInjectedViewer()
-const getPreviewUrl = useGetPreviewUrl()
-const { resourceItems } = useResolvedViewerResources()
-
-const author = computed(() => props.version.authorUser)
-
-const createdAt = computed(() =>
-  dayjs(props.version.createdAt as string).format('DD MMM YY, h:mm A')
-)
-
-const timeAgoCreatedAt = computed(() =>
-  dayjs(props.version.createdAt as string).from(dayjs())
-)
-
-const previewUrl = computed(() =>
-  getPreviewUrl(projectId.value, props.version.referencedObject)
-)
-
-function handleClick() {
-  if (props.clickable) emit('change-version', props.version.id)
-}
-</script>
